Throw error when supplied targetUrl to switchTo is empty #1057 (#1118)
* Throw error when supplied targetUrl to switchTo is empty #1057
* Checked for only spaces #1057
* Remove period

The full message is "Error: Cannot switch to tab or window. Hint: The targetUrl is empty. Please use a valid string or regex, run `.trace` for more info."

* Update test/unit-tests/switchTo.test.js

The full message is "Error: Cannot switch to tab or window. Hint: The targetUrl is empty. Please use a valid string or regex, run `.trace` for more info."

* Update test/unit-tests/switchTo.test.js

The entire message is "Error: Cannot switch to tab or window. Hint: The targetUrl is empty. Please use a valid string or regex, run `.trace` for more info."

Co-authored-by: Vinay Shankar Shukla <vinaysh@thoughtworks.com>
Co-authored-by: Zabil Cheriya Maliackal <zabil@users.noreply.github.com>