Test types for browser actions (#1462)
* Added type test from specs for openBrowser, closeBrowser

The commit adds some tests for the functions openBrowser and closeBrowser.
The tests are based on the examples provided in the official documentation.

Signed-off-by: Filippo Maria Panelli <filippo.panelli@gmail.com>

* Added type test for switchTo and fixed switchTo type def

This commit adds tests for switchTo type definition. The tests are
derived from the examples in the official documentation.

Adding those tests shows that the type definition was not correct.
This commit fixes the types of the arguments of switchTo function.

Signed-off-by: Filippo Maria Panelli <filippo.panelli@gmail.com>

* Added type test for intercept and fixed type def

This commit adds tests for intercept type definition. The tests are
derived from the examples in the official documentation and from
Nivedha's comment here
https://github.com/getgauge/taiko/issues/98#issuecomment-420241863.

Adding those tests shows that the type definition was not correct.
This commit fixes the types of the arguments of intercept function.

Signed-off-by: Filippo Maria Panelli <filippo.panelli@gmail.com>

* Added type tests for emulateNetwork, emulateDevice, setViewport

Signed-off-by: Filippo Maria Panelli <filippo.panelli@gmail.com>

* Added type test for openTab and fixed type def

This commit adds tests for openTab type definition. The tests are
derived from the examples in the official documentation.

Adding those tests shows that the type definition was not correct.
This commit fixes the types of the arguments of openTab function.

Signed-off-by: Filippo Maria Panelli <filippo.panelli@gmail.com>

* Added type test for closeTab and fixed type def

This commit adds tests for closeTab type definition. The tests are
derived from the examples in the official documentation.

Adding those tests shows that the type definition was not correct.
This commit fixes the types of the arguments of openTab function.

Signed-off-by: Filippo Maria Panelli <filippo.panelli@gmail.com>

* Added type test for several taiko functions

This commit adds tests for type definitions for:

- openIncognitoWindow
- closeIncognitoWindow
- overridePermissions,
- clearPermissionOverrides

The tests are derived from the examples in the official documentation.

Signed-off-by: Filippo Maria Panelli <filippo.panelli@gmail.com>

* Added type tests for the remaining browser actions

This commit adds tests for type definitions for:

- setCookie
- deleteCookies
- getCookies
- setLocation
- clearIntercept

The tests are derived from the examples in the official documentation.

This commit also introduces the concept of getting some types from
the official devtools-protocol repository.

Signed-off-by: Filippo Maria Panelli <filippo.panelli@gmail.com>

Co-authored-by: Nivedha <nivedhasenthil@gmail.com>